[WEB-3828] - clean up for SoC
