+++ src/ecanvasserIntegration/ecanvasserIntegration.types.ts
@@ -1,23 +1,3 @@
-import { Prisma } from '@prisma/client'
-
-export type EcanvasserWithRelations = Prisma.EcanvasserGetPayload<{
-  include: {
-    contacts: true
-    houses: true
-    interactions: true
-    campaign: {
-      select: {
-        id: true
-        user: {
-          select: {
-            email: true
-          }
-        }
-      }
-    }
-  }
-}>
-
 export interface EcanvasserSummary {
   contacts: number
   houses: number
@@ -26,64 +6,6 @@ export interface EcanvasserSummary {
   campaignId: number | undefined
   lastSync: Date | null
   error: string | null
-}
-
-export interface EcanvasserContact {
-  id: number
-  first_name: string
-  last_name: string
-  type: string
-  gender?: string | null
-  date_of_birth?: string | null
-  year_of_birth?: number | null
-  house_id?: number | null
-  unique_identifier?: string | null
-  organization?: string | null
-  volunteer: boolean
-  deceased: boolean
-  donor: boolean
-  contact_details: {
-    email: string | null
-    home: string | null
-    mobile: string | null
-  }
-  action_id?: number | null
-  last_interaction_id?: number | null
-  created_by: number
-  created_at: string
-  updated_at: string
-  customFields?: Array<any>
-}
-
-export interface EcanvasserInteraction {
-  id: number
-  rating?: number | null
-  status: {
-    id: number
-    name: string
-    description: string
-    color: string
-  }
-  effort_id: number
-  contact_id?: number | null
-  house_id?: number | null
-  type: string
-  action_id?: number | null
-  created_by: number
-  created_at: string
-  updated_at: string
-  survey?: {
-    id: number
-    name: string
-    description: string
-    status: string
-    created_by: number
-    updated_at: string
-    created_at: string
-    team_id: number | null
-    requires_signature: boolean
-    responses: any[]
-  } | null
 }
 
 // API Response Types (snake_case)
@@ -120,25 +42,6 @@ export interface ApiEcanvasserInteraction {
   contact_id?: number
   created_by: number
   created_at: string
-}
-
-export interface InteractionsByDay {
-  [date: string]: {
-    count: number
-    [status: string]: number
-  }
-}
-
-export interface EcanvasserSummaryResponse {
-  totalContacts: number
-  totalHouses: number
-  totalInteractions: number
-  averageRating: number
-  interactions: {
-    [key: string]: number
-  }
-  interactionsByDay: InteractionsByDay
-  lastSync: Date | null
 }
 
 export interface PaginationParams {
@@ -196,6 +99,30 @@ export interface ApiEcanvasserSurveyQuestion {
   answers?: Array<{
     name: string
   }>
+}
+
+export interface ApiEcanvasserHouse {
+  id: number
+  unit: string
+  number: string
+  name: string
+  address: string
+  city: string
+  state: string
+  latitude: number
+  longitude: number
+  source: string
+  location_type: string
+  last_interaction_id: number
+  action_id: number | null
+  building_id: number | null
+  type: string
+  zip_code: string
+  precinct: string
+  notes: string
+  created_by: number
+  created_at: string
+  updated_at: string
 }
 
 export interface ApiEcanvasserTeam {
